make static build compatible with cmake
Common cmake script uses builtin `FindZLIB` module to find where zlib is
installed. The cmake crate allows `register_dep` to hint cmake to search
the right place. libz-sys puts the library under build directory, which
is however not the place where cmake is going to search. So this PR
corrects the behavior by moving the built library to lib directory.

Another option is to set the root directory to build instead, but the
layout may not look good.

It has been an issue for downstream projects like tikv/rust-rocksdb#303.
and tikv/grpc-rs#419.