Allow building zlib without libc
Adds a `libc` feature which is enabled by default, but opting out from it allows to build zlib with Z_SOLO mode.

This mode removes dependency on any external libraries including libc at the cost of eliminating some high-level functions like gz*, compress* and uncompress, and requiring embedder to provide memory allocation routines to deflate and inflate.

This is useful for targets without libc support such as embedded targets or wasm32-unknown-unknown, as well as for embedding in crates that provide own allocation anyway.

An an example of the latter, when using this mode from a local branch of flate2, the output artifact of `cargo build --release --example compress_file --features zlib` in flate2 goes down by 7KB, even though all the functionality is still available as before.

Note: the reason we're introducing a feature `libc` and not a more natural variant `solo` is because Cargo takes features from all dependencies into the account. That way, if we had at least one dependency that uses libz-sys with `solo` feature, then all other dependencies would also get a Z_SOLO build, which might break their expectations if they depend on other high-level functions. Instead, by providing a `libc` feature, we guarantee that we build in Z_SOLO mode only if none of the dependencies needs the full zlib. For that, all such dependencies must use `default-features = false`.

This build also removes a dependency on the Rust libc crate, because otherwise we would need to disable it conditionally per-target, even though we only need types available in standard library under `std::os::raw` and using them allows to get rid of one more dependency altogether.