Add a libz-ng-sys crate, using the same sources with minimal duplication
libz-ng-sys uses the zlib-ng native API, rather than the zlib-compat
API. It exports the same Rust API (modulo the slight differences in
types), without the `zng_` prefixes, to make it easy to write Rust
software that works with both.

Add a separate Cargo.toml and README for libz-ng-sys.

Move the `build_zlib_ng` function from `build.rs` to `build_zng.rs`, and
use that as the build script for libz-ng-sys.

Add a cargo-zng script that creates the libz-ng-sys crate in a temporary
directory and runs cargo on it, to make it easy to run tests with:
```
./cargo-zng test
./cargo-zng run --manifest-path systest/Cargo.toml
```

Test libz-ng-sys in CI.