Fix zlib_installed not using the correct flags
The `zlib_installed` function does not use the correct compiler flags
when checking if there already is a libz library pre-build.

This causes libz-sys to build the internal libz version, which might be
different from the pre-build libz version, and might cause issues with
other libraries which also used the pre-build libz wether or not it's
static or shared.

I came across this when using my custom toolchain build images to build
static musl binaries. And this crate seemed to build the built-in
version anyways. Which is an other issue, not addressed in this PR.