Use `RISCV_WITH_RVV` as the env var, taking ON/OFF values
Instead of `RVV_OFF`.

This checks if the value appears to match any of several hard-coded
truthy or falsy values. Typically the literal values `ON` or `OFF`
would be used, as is typically done when passing boolean `-D...`
options to `cmake` on the command-line. It is typically more useful
to pass `OFF` here than `ON`, becuase the default is `ON`, as noted
in the linked section of the zlib-ng readme.

At least for now, empty, blank, and otherwise unrecognized values
are simply ignored. Recognized values are always "normalized" as
either `ON` (if truthy) or `OFF` (if falsy).